requirements/default.txt
django-auth-kerberos==1.2.4